+++ mayan/apps/cabinets/search.py
@@ -5,7 +5,7 @@ from mayan.apps.dynamic_search.classes import SearchModel
 from .permissions import permission_cabinet_view
 
 cabinet_search = SearchModel(
-    app_label='cabinets', model_name='Cabinet',
+    app_label='cabinets', model_name='CabinetSearchResult',
     permission=permission_cabinet_view,
     serializer_path='mayan.apps.cabinets.serializers.CabinetSerializer'
 )